chore(SR4): update central api checkApiKey to be aligned with external-api checkApiKey
